refactoring: change PlayNext folder to PlayNextView
